==> JupyterNotebooks/02_DataProcessing.ipynb <==
<p><span style="font-size:36px"><strong>2 Data Processing</strong></span></p>

import pandas as pd
import geopandas as gpd
import json as js
import numpy as np
import topojson as tp
import datetime
import shapely
from shapely import wkt
from shapely.wkb import loads, dumps
from shapely.geometry import Polygon, LineString, Point

<p><span style="font-size:18px"><strong>2.1 Elektroladestationen JSON in DataFrame umwandeln</strong></span></p>

Quelle: https://opendata.swiss/de/dataset/ladestationen-fuer-elektroautos/resource/4d467a51-0bc9-48ce-aa2a-3d3bcaa7e7e9

#JSON wird geladen
f = open('../Data/lade.json')
data = js.load(f)

#JSON wird in DF umgewandelt
OperatorNames = []
cities = []
plz = [] 
coords = []
for name in data['EVSEData']:
    for add in name['EVSEDataRecord']:
        OperatorNames.append(name['OperatorName'])
        cities.append(add['Address']['City'])
        plz.append(add['Address']['PostalCode'])
        coords.append(add['GeoCoordinates']['Google'])
        
d = {'OperatorName':OperatorNames,'City':cities,'Plz':plz,'Coord':coords}
df = pd.DataFrame(data=d)

#Koordinaten werden für spätere Schritte in zwei Spalten gesplittet
df['Lat']= df.apply(lambda x: x['Coord'].split(' ')[0], axis=1)
df['Lon']= df.apply(lambda x: x['Coord'].split(' ')[1],axis =1)

<p><span style="font-size:18px"><strong>2.2 Gemeindegrenzen richtig formattieren</strong></span></p>

Die Gemeindegrenzen wurden in einem vorherigen Schritt (siehe DataCollection Notebook) heruntergeladen. Jetzt gilt es die Daten in ein gängiges Format umzuwandeln, mit welchem die Grenzen richtig ausgelesen werden können. In dieser Domäne bietet sich ein GeoDF von GeoPandas sehr gut an, da GeoPandas viele Funktionen zum handeln von geografischen Daten besitzt.

gemdf = pd.read_csv('../Data/GemeindeGrenzen.csv')

Da Geopandas eine spezielle Formatierung braucht um die GeoDaten(MultiPolygone) als solche zu erkennen und es uns nicht gelungen ist diese korrekt einzulesen, werden Sie hier manuell in die korrekte Formatierung gebracht.

polytpl = []
for place in gemdf.Polygon:
    lst1 = [] 
    place = js.loads(str(place))['coordinates']
    for polobj in place:
        lst2 = []    
        for xobj in polobj:
            lst3 = []    
            for coordinates in xobj:
                   
                container = ''
                counter = 0
                for latlan in coordinates:
                    if counter == 1:
                        container += ' '
                    container += str(latlan)
                    counter +=1  
                lst3.append(container)
            lst2.append(tuple(lst3))
        lst1.append(tuple(lst2))
    polytpl.append(tuple(lst1))
polytpl = tuple((polytpl))

#Formatierung wird abgespeichert
gemdf['FormatPolygon'] = polytpl

#Daten werden weiter korrekt formatiert
gemdf['geometry'] = gemdf.apply(lambda x: str('MULTIPOLYGON' + str(x['FormatPolygon'])), axis = 1 )

#Daten werden zu ende formatiert und werden eingelesen
geometry_objects = [shapely.wkt.loads(str(s).replace("'",'').replace("),)",'))').replace(")),)",')))')) for s in gemdf.geometry]

#Daten werden als GeoPanda DF abgespeichert
geodf = gpd.GeoDataFrame(gemdf[['BFSid','GemeindeLabel']], geometry=gpd.GeoSeries(geometry_objects))

#Kontrolle
geodf

<p><span style="font-size:18px"><strong>2.3 Anzahl Elektroladestationen pro Gemeinde bestimmen</strong></span></p>

#DF wird auf korrektheit kontrolliert
df[df['OperatorName']=='eCarUp'].sort_values(by='Lat',ascending=True).head(60)

Wie man sehen kann gibt es gewisse Betreiber mit fehlerhafte Ladestationendaten, welche nicht in der Schweiz liegen. Zum Glück stimmen alle Koordinaten, weswegen man diese mit den Gemeindegrenzen abgleichen und die korrekte Gemeinde zuweisen kann.

#Koordinaten werden auf die Gemeinden abgegliechen und Gemeinde herausgesucht
ladeort = []
for x in range (0,df.shape[0]):
    try:
        ladeort.append(geodf[geodf.contains(Point(df['Lon'].iloc[x],df['Lat'].iloc[x]))]['GemeindeLabel'].values[0])
    except:
        ladeort.append(str('Error'))

#Die korrekten Gemeindebezeichnungen werden hinzugefügt
df['GemeindeLoc'] = ladeort

#Überblick über die falschen Daten und 'problematische' Betreiber, welche fehlerhafte Daten haben
df[(df['GemeindeLoc']=='Error') & (df['OperatorName']!='eCarUp')]

#Der DF wird mit den korrekten Gemeinde IDs vom Bund (BFS-ID) versehen
BFSidlist = []
for x in range(0,df.shape[0]):
    try:
        BFSidlist.append(gemdf[gemdf['GemeindeLabel'] == df.GemeindeLoc.iloc[x]].BFSid.values[0])
    except:
        BFSidlist.append('NAN')

#Die BFS-ID werden hinzugefügt
df['BFSid'] = BFSidlist

#Das DF wird gecleaned. Nicht vorhandene Gemeinden (z. bsp. welche nicht in der Schweiz liegen werden gedroppt)
df = df.drop(df[df['BFSid'] == 'NAN'].index, axis=0)
df = df.drop(columns=['City','Plz','Coord'], axis=1)

geodf = geodf.merge(df.groupby('BFSid').count()['OperatorName'],how='left',on='BFSid')

#Ladestationen pro Gemeinde werden gezählt
geodf['AnzahlLadestation'] = geodf['OperatorName']
geodf['AnzahlLadestation'] = geodf['AnzahlLadestation'].fillna(0)
geodf = geodf.set_index('BFSid')
geodf = geodf.drop(columns='OperatorName', axis =1)

geodf.to_csv('../Data/LadestationProGemeinde.csv')

<p><span style="font-size:18px"><strong>2.4 Geodaten simplifizieren um Speicher/Leistung zu sparen </strong></span></p>

Da die Gemeindedaten sehr speicher intensiv sind und auf Streamlit mehr performance zum laden brauchen werden sie in den unteren Schritten vereinfacht. Die Vereinfachungsfunktion von GeoPandas vereinfacht leider nicht ohne dass Lücken zwischen den Gemeinden entstehen, so dass auf die topojson liabrary ausgewichen werden musste.

geodf_simple = tp.Topology(geodf, topology=True, toposimplify=0.001).to_gdf()

geodf_simple.to_csv('../Data/GemeindeGrenzen_Vereinfacht.csv')

<p><span style="font-size:18px"><strong>2.5 Ladestation Timeseries vorbereiten</strong></span></p>

Für die TimeSeries in unserer StreamlitApp bereiten wir hier die Daten in einer sogenannten 'geschmolzenen' Tabelle vor.<br /> Quelle: https://opendata.swiss/de/dataset/kennzahlen-offentliche-ladeinfrastruktur-elektromobilitat/resource/d772c474-e890-4de6-b1bb-416d90e860a3

ladstat = pd.read_csv('../Data/charging-infrastructure_key-figures.csv')

#Überblick Kolumnen
ladstat.columns.values

#Alle Kolumnen welche 'stations' enthalten sind für uns interessant. 
#Deswegen erstellen wir eine Liste mit diesen Kolumnen, welche wir nutzen werden um den DF zu filtern.
subs = 'stations'
res = [i for i in ladstat.columns.values if subs in i]
res.append('year')
res.append('month')
res

#Das gefilterte DF wird 'geschmolzen'
melted_df = ladstat[res].melt(id_vars=['year','month'],var_name='Canton',value_name='Amount')

#DF wird gecleaned
melted_df['Canton'] = melted_df.apply(lambda x: x['Canton'].replace('stations_','').replace('_count',''),axis=1)
melted_df['day'] = 1
melted_df['datetime'] = pd.to_datetime(melted_df[['year','month','day']])
melted_df = melted_df.drop(columns =['year','month','day'],axis=1)

#Kontrolle
melted_df

melted_df.to_csv('../Data/Chargingstations_melted.csv')
