==> JupyterNotebooks/03_DataAggrigation.ipynb <==
# 3. Aggregierung der Daten

Die meisten der genutzten Daten sind auf dem Stand vom Jahr 2021. Die Standorte per Gemeinde sind jedoch auf dem neusten Stand. Im Jahr 2022 wurden einige Gemeinden miteinander fusioniert. Des Weiteren wiesen gewisse Gemeindedaten Lücken auf, welche sich durch eine einfache Internet-Recherche füllen liessen.

All diese Daten wurden direkt in den Excel Files ergänzt und sind daher im Code nicht ersichtlich.

import pandas as pd

Die historischen Daten der Elektorautos die einglöst sind auf Gemeinde-Level werden importiert und die BFS-Nr. wird als Index definiert.

ElektroautoBestand = pd.read_excel(
    '00_data\Bestand_Elektrofahrzeuge_Gemeinden_2010-2021.xlsx',
    names=[
        'Gemeinde',
        'EV_Bestand_2010', 
        'EV_Bestand_2011', 
        'EV_Bestand_2012', 
        'EV_Bestand_2013', 
        'EV_Bestand_2014', 
        'EV_Bestand_2015', 
        'EV_Bestand_2016', 
        'EV_Bestand_2017', 
        'EV_Bestand_2018', 
        'EV_Bestand_2019', 
        'EV_Bestand_2020', 
        'EV_Bestand_2021'
        ],
    header=0
    )

ElektroautoBestand[['BFS-Nr','Gemeinde']] = ElektroautoBestand["Gemeinde"].str.split(" ", 1, expand=True)
ElektroautoBestand['BFS-Nr'] = ElektroautoBestand['BFS-Nr'].astype('int64')
ElektroautoBestand.set_index('BFS-Nr', inplace=True)
ElektroautoBestand.drop('Gemeinde', axis=1, inplace=True)
ElektroautoBestand

Die Strassenlängen der einzelnen Gemeinden (Stand 2021) werden importiert und die BFS-Nr. wird als Index definiert.

Strassenlänge = pd.read_excel(
    '00_data\Strassenlänge_nach_Gemeinde.xlsx',
    names=[
        'Kantons-Nr', 
        'Kanton', 
        'BFS-Nr', 
        'Gemeinde', 
        'Strassenlänge(km)']
    )

Strassenlänge.drop(['Kantons-Nr', 'Kanton', 'Gemeinde'], axis=1, inplace=True)
Strassenlänge['BFS-Nr'].astype('int64')
Strassenlänge.set_index('BFS-Nr', inplace=True)
Strassenlänge

Die einzelnen Gemeindeerhebungen der Bevölkerung werden importiert und die BFS-Nr. wird als Index definiert.

Census = pd.read_excel(
    '00_data\Regionalporträts_Gemeinden.xlsx',
    names=[
        'BFS-Nr',
        'Gemeindename',
        'Anz_Einwohner',
        'Einwohner_Veränderung_2010-2019',
        'Bevölkerungsdichte_pro_km2',
        'Ausländeranteil',
        'Altersverteilung_0-19',
        'Altersverteilung_20-64',
        'Altersverteilung_65+',
        'Rohe_Heiratsziffer',
        'Rohe_Scheidungsziffer',
        'Rohe_Geburtenziffer',
        'Rohe_Sterbeziffer',
        'Anz_Privathaushalte',
        'Durchschn_Haushaltsgrösse',
        'Gesamtfläche(km2)',
        'Siedlungsfläche',
        'Siedlungsfläche_Veränderung(ha)',
        'Landwirtschaftsfläche',
        'Landwirschaftsfläche_Veränderung(ha)',
        'Wald&Gehölze',
        'Unproduktive_Fläche',
        'Beschäftigte_total',
        'Beschäftigte_1_Sektor',
        'Beschäftigte_2_Sektor',
        'Beschäftigte_3_Sektor',
        'Arbeitsstätte_total',
        'Arbeitsstätten_1_Sektor',
        'Arbeitsstätten_2_Sektor',
        'Arbeitsstätten_3_Sektor',
        'Leerwohnungsziffer',
        'Neu_gebaute_Wohnungen(pro_1000_Einwohner)',
        'Sozialhilfequote',
        'Wahlanteil_FDP',
        'Wahlanteil_CVP',
        'Wahlanteil_SP',
        'Wahlanteil_SVP',
        'Wahlanteil_EVP',
        'Wahlanteil_GLP',
        'Wahlanteil_BDP',
        'Wahlanteil_PdA/Sol.',
        'Wahlanteil_GPS',
        'Wahlanteil_kleine_Rechtsparteien',
    ]
)

Census.drop('Gemeindename', axis=1, inplace=True)

Census['BFS-Nr'].astype('int64')
Census.set_index('BFS-Nr', inplace=True)

Census

Die Abstimmungsresultate über die CO2-Vorlage werden importiert und die BFS-Nr. wird als Index definiert.

Abstimmung = pd.read_excel(
    "00_data\co2-gesetz-abstimmung_nach_gemeinde.xlsx",
    names= [
        'BFS-Nr', 
        'Gemeindename', 
        'CO2-Gesetz_Ja-Anteil', 
        'EidgVorl_4_winning_formula', 
        'CO2-Gesetz_Stimmbeteiligung'
        ],
    header=0
    )

Abstimmung.drop(['Gemeindename', 'EidgVorl_4_winning_formula'], axis=1, inplace=True)

Abstimmung['BFS-Nr'].astype('int64')
Abstimmung.set_index('BFS-Nr', inplace=True)
Abstimmung

Die zuvor webgescrapten Ladestationen pro Gemeinde werden importiert und die BFS-Nr. wird als Index definiert.

Ladestationen = pd.read_csv(
    "00_data\LadestationProGemeinde.csv",
    names=[
        'BFS-Nr',
        'Gemeindename',
        'GeoLocation',
        'Ladestationen'
        ],
    header=0
    )

Ladestationen['BFS-Nr'].astype('int64')
Ladestationen.set_index('BFS-Nr', inplace=True)

Ladestationen

Alle Datasets werden zu einem gesamthaften Datensatz auf Basis der BFS-Nr. gemerged.

Die BFS-Nr. wurde generell überall als Index genommen, da diese konsistent über alle Datasets ist und demnach einfach zu handhaben ist.

data = Ladestationen.merge(ElektroautoBestand, how='left', left_index=True, right_index=True)
data = data.merge(Strassenlänge, how='left', left_index=True, right_index=True)
data = data.merge(Census, how='left', left_index=True, right_index=True)
data = data.merge(Abstimmung, how='left', left_index=True, right_index=True)
data

nan_data = data[data.isna().any(axis=1)]
nan_data

Die einzigen zwei Instanzen, welche keine Daten aufweisen, sind eigentlich keine politischen Gemeinden, sondern ein Gemeinschaftsareal mehrer Gemeinden.
Aus diesem Grund wurde entschieden, diese Instanze nicht zu berücksichtigen

data.dropna(inplace=True)
data
data.to_csv('aggrigated_data.csv')
